+++ src/Ext/Techno/Body.Internal.cpp
@@ -148,6 +148,29 @@ CoordStruct TechnoExt::GetSimpleFLH(InfantryClass* pThis, int weaponIndex, bool&
 	return FLH;
 }
 
+void TechnoExt::ExtData::InitializeUnitIdleAction()
+{
+	TechnoClass* const pThis = this->OwnerObject();
+
+	if (pThis->WhatAmI() != AbstractType::Unit || !pThis->HasTurret())
+		return;
+
+	TechnoTypeClass* const pType = pThis->GetTechnoType();
+	TechnoTypeExt::ExtData* const pTypeExt = this->TypeExtData;
+
+	if (pTypeExt->AutoFire || pType->TurretSpins)
+		return;
+
+	if (pTypeExt->UnitIdleRotateTurret.Get(RulesExt::Global()->UnitIdleRotateTurret))
+		this->UnitIdleAction = true;
+
+	if (!SessionClass::IsSingleplayer())
+		return;
+
+	if (pTypeExt->UnitIdlePointToMouse.Get(RulesExt::Global()->UnitIdlePointToMouse))
+		this->UnitIdleActionSelected = true;
+}
+
 void TechnoExt::ExtData::InitializeAttachEffects()
 {
 	if (auto pTypeExt = this->TypeExtData)